Make felm statistics for projected model optional (#157)
* Make felm statistics for projected model optional

The extract method for the lfe package includes statistics (e.g.
R-squared) for the projected model in the GOF block.

This commit adds a parameter to extract.felm that allows users to
exclude statistics for the projected model. The default behavior,
however, is unchanged.

* Test exclusion of projected model stats

Test that user can optionally exclude summary statistics for the
projected model when using felm from the lfe package.